==> 2_data_prep.Rmd <==
---
title: "Prep input data, including species cleaning"
author: "GGM & RR"
date: "09/29/2021"
output: html_notebook
---

## Set paths
```{r}
# delphi_prosjekt_drive <- "U:" # which letter have you mapped this drive to?
# 
# dataPath <- file.path(delphi_prosjekt_drive,"Mareano/VIDEOLAB/VIDEO DATA/200m_scale_species_by_sample/Data_Delivery_2021")
# outPath <- file.path(dataPath,"outputs")
# codePath <- getwd()
```

## Libraries
```{r}
library(readxl)
library(dplyr)
```

## Read data
```{r}
spmatrix <- read.csv(file.path(dataPath,"species_densities.csv"), sep ="|")
```

## Edit species
### import edits file
```{r}
edits <- read_xlsx(file.path(dataPath,"Further editing of taxa for analyses.xlsx"))
edits <- mutate(edits, original_name = gsub("&",";",`Row Labels`))
head(edits)
```

### modify species according to the edits
```{r}
# transcode species
spmatrix <- spmatrix %>% mutate(new_name = edits$`Replace with`[match(spmatrix$clean_taxonomy,edits$original_name)])
idx <- which(is.na(spmatrix$new_name))
spmatrix$new_name[idx]<- as.character(spmatrix$clean_taxonomy[idx])

# replace Baltina sp with Halipteris
#spmatrix <- spmatrix %>% mutate(new_name = gsub("Balticina sp.",
#                                                  "Halipteris sp.",
#                                                  new_name))

# Gracilechinusacutus is split from Gracilechinusacutus;juvenile should we keep split?
spmatrix <- spmatrix %>% mutate(new_name = gsub("Gracilechinus acutus;juvenile",
                                                  "Gracilechinus acutus",
                                                  new_name))

# ?	Suspect tunicatatrunk and ascidiaceasolitarybig are the same (could leave split, Pål to decide?)
# ?	Should we remove buccinidae?
# ?	Should we remove paralithodescamtschaticus? (they migrate long distances so are not associated with any particular habitat)
# I think we should merge Pandalus and pandalidae

# remove species
editrem <- edits %>% subset(Delete == "x")
idx <- which(is.na(match(spmatrix$clean_taxonomy,editrem$original_name)))
#idx <- which(is.na(match(spmatrix$new_name,editrem$original_name))) #or this?

# trimmed matrix
trimmed <- spmatrix[idx,]
```


## Merge samples when possible
When looking at KF I realized that some samples were falling out of the analysis but could be merged beacuse they were paired with another one. I should look for other cases like this more systematically...
```{r}
# 2120_00 and 2121_B (R station 2065) Second line was taken because of a crash during the first one. data can be combined

trimmed <- trimmed %>% mutate(SampID = gsub("2120_00", "2121_B", SampID)) %>%
  group_by(., SampID, new_name) %>%
  summarise(., TotAbu_pseudocount = sum(TotAbu_pseudocount),
            TotAbu_count = sum(TotAbu_count),
            density_n_100m2_pseudocount = sum(density_n_100m2_pseudocount),
            density_n_100m2_count = sum(density_n_100m2_count))

# sample_info$SegLengthM <- with(sample_info, replace(SegLengthM, which(SampID == "2121_B"), sum(57.4500,114.4500)))
# sample_info$area_m2 <- with(sample_info, replace(area_m2, which(SampID == "2121_B"), sum(123.1584,413.2917)))
# sample_info$NumSplits <- with(sample_info, replace(NumSplits, which(SampID == "2121_B"), -1))

# 1920_00 and 1920_01 (R station 1861) I honestly don't know why this was split in the first place. data can be combined

trimmed <- trimmed %>% mutate(SampID = gsub("1920_01", "1920_00", SampID)) %>%
  group_by(., SampID, new_name) %>%
  summarise(., TotAbu_pseudocount = sum(TotAbu_pseudocount),
            TotAbu_count = sum(TotAbu_count),
            density_n_100m2_pseudocount = sum(density_n_100m2_pseudocount),
            density_n_100m2_count = sum(density_n_100m2_count))

# sample_info$SegLengthM <- with(sample_info, replace(SegLengthM, which(SampID == "1920_00"), sum(135.5369,135.5369)))
# sample_info$area_m2 <- with(sample_info, replace(area_m2, which(SampID == "1920_00"), sum(210.6630,236.5845)))
# sample_info$NumSplits <- with(sample_info, replace(NumSplits, which(SampID == "1920_00"), -1))
# sample_info$ok_vision <- with(sample_info, replace(ok_vision, which(SampID == "1920_00"), mean(0.2235294,0.6363636)))
# sample_info$poor_vision <- with(sample_info, replace(poor_vision, which(SampID == "1920_00"), mean(0.7764706,0.3636364)))


```

## Result
```{r}
unique(data.frame(sort(trimmed$new_name)))
write.csv(trimmed,file.path(outPath,"species_densities_trimmed.csv"),row.names = FALSE)
```
Output metadata to have list of included samples and their locations for mapping checks (are any obviously missing?)
```{r}
start.dta<-trimmed %>%
  select(SampID)%>%
  distinct() %>%
  left_join(sample_info)

write.csv(start.dta, file = file.path(outPath, ("start_dta.csv")))
```
