==> 2b_data_prep_envVarPerSamp.Rmd <==
---
title: "Prep env var per sample - R Notebook"
output: html_notebook
---

#Libraries

```{r}
library(dplyr)
library(tidyr)
library(raster)
```
#Load sample info
also loaded in 3_data_analysis_twinspan, but only extracting locations here
```{r}
##uncomment if not already loaded
sample_info_raw<-read.csv(file.path(dataPath,"sample_info.csv"))

samp_pos<-sample_info_raw %>%
  dplyr::select(SampID, x_coord, y_coord)


#make shapefile for sampling rasters
samp.shp <- SpatialPointsDataFrame(coords = cbind(samp_pos$x_coord,
                                                  samp_pos$y_coord),
                                 proj4string = CRS("+init=epsg:32633"),
                                 data =  samp_pos)

```

#Load video log extracted depth and sediment data
Depth
```{r}
samp_d<-read.csv(file.path(dataPath,"depthMeans.csv"))
``` 
 
Sediment 
```{r} 
samp_sed<-read.csv(file.path(dataPath,"sedimMeans.csv"))

samp_sed <- samp_sed[-which(samp_sed$Substrate == ""), ]
#samp_sed$Substrate<-sub("^$", "Unknown", samp_sed$Substrate)


samp_sed_wide<-samp_sed %>%
  pivot_wider(id_cols = SampID,
              names_from = Substrate,
              values_from = mean_percent,
              values_fill = NA)
```



#Load any rasters to sample

```{r}
oceanogPath<-"U:/Mareano/OCEANOGRAPHY/JonAlbretsenData"
oceanog.files.names<-list.files(oceanogPath)
oceanog.files<-list.files(oceanogPath,full.names = T)
oceanog.stk<-stack(oceanog.files)
names(oceanog.stk)<-unlist(oceanog.files.names)
```


#sample rasters and convert to dataframe
```{r}
samp.env.shp<-extract(oceanog.stk, samp.shp, sp=T)
samp.env.o<-as.data.frame(samp.env.shp)%>%
  dplyr::select(-c("coords.x1","coords.x2"))
```

#assemble sample env var lookup table and write out

```{r}
samp.envVar<-samp.env.o %>%
  left_join(samp_d)%>%
  left_join(samp_sed_wide)

write.csv(samp.envVar, file = file.path(outPath, ("samp_envVar.csv")))
```
