==> 3_KR_kongsRijpTWINSPAN.Rmd <==
---
title: "KongsRijp TWINSPAN - R Notebook"
output: html_notebook
---

Assumes you've been working with the megamatrix/TWINSPAN first as subsetting from that dataset.
## libraries
```{r}
library(twinspan)
library(dplyr)
library(RColorBrewer)
library(heatmaply)
library(tibble)

```


## SUB TWINSPAN
e.g. kongsfjorden/rijpfjorden set
```{r}
sites<-sample_info %>%
  filter(VL %in% fjord$sample_no) %>%
  select(SampID)#110 samples

sub.coredata<-coredata %>%
  filter(row.names(coredata) %in% sites$SampID)%>%#69 samples retained
  filter(!(row.names(.) %in% c("2108_B","2109_B","2110_B","2111_B","2113_B",
                                     "2114_B", "2115_B","2121_B","2122_B","2123_B")))

tw_kr <- sub.coredata%>%
  select(which(!is.na(colSums(.)))) %>%
  #mutate_all(sqrt)%>%
  twinspan(cutlevels = cutlevelsPaal, levmax = 15, groupmin = 3)
```

## check mega dataset for lost samples
search for where they were removed and store the list. 

```{r}
# #list of samples to search that are in mega analysis dataset
# core.samps<-coredata %>%
#   rownames_to_column(var="SampID")%>%
#   select(SampID)
# 
# view(core.samps)
# #list from mega analysis of excluded samples
# 
# view(res.missing)
# 
# #list species that were na colSums (does this filter out some stations?)
# no.selspp<-coredata %>%
#   select(which(is.na(colSums(.)))) #? looks like these just had some vals as text
# 
# #check refer (the first loading in of data that came from MarVid)
# fj.refer<-refer %>%
#   filter((sample_no %in% fjord$sample_no))

#check good
fj.bad<-fjord %>%
  filter(!(sample_no%in%good))

fj.bad

# #check new filter
# tw.newfilt <- trimmedwide %>%
#   data.frame() %>%
#   filter(new_filter)

#see reasons for newfilt for fjord stations
fj.newfilt.out<-trimmedwide %>%
  data.frame() %>%
  filter(!(new_filter))%>%
  select(SampID)%>%
  left_join(sample_info)%>%
  filter(VL %in% fjord$sample_no)
fj.newfilt.out

```
Looks like:
- 3 samples did not make it into the "good" filter i.e. the total number of observations/the total length of the dive was <5 (NB one of these was also too short 69m due to dive being aborted: 1903)
- 39 samples were filtered out by "new filter" i.e. they had poor vision for >0.75 of the dive (there are 38 listed, one is in the 3 above (1903) and two (1920 and 1929) have two samples each i.e. 1920_00 and 1920_01, 1929_00 and 1929_01)
=41 samples which are all that are missing.
Later we will try and predict these into TWINSPAN classes at a level that is deemed relevant for biotope classes.


## twintable of selected sites subset
best to just copy and paste output from screen to text file - does not easily capture with automated methods
```{r}

twintable(tw_kr, goodspecies = "both")

# pdf(file.path(outPath,"twintableKR.pdf"), width=10, height=40)
# twintable(tw_kr, goodspecies = "both")
# dev.off()
# 
# 
# capture.output(twintable(tw_kr, goodspecies = "both"), file = file.path(outPath,"twintableKR.pdf"), append = FALSE,
#                type = c("output", "message"), split = FALSE)
```

## dendrograms
```{r}
#MINI WITH LABELLED NODES AND GROUPS

plot(tw_kr, "quadrat")
plot(tw_kr,"species")

# pdfs

pdf(file.path(outPath,"miniDendro_tw_kr.pdf"), width=20, height=10) #adjust width as necessary
par(cex=1)
plot(tw_kr,"quadrat")
dev.off()

pdf(file.path(outPath,"miniDendro_tw_kr_spp.pdf"), width=20, height=10)
plot(tw_kr,"species")
dev.off()


# TRIANGLE QUADRAT OR SPECIES LABELS
plot(as.dendrogram(tw_kr,"species"),type="triangle")

#pdfs

pdf(file.path(outPath,"Dendro_tw_kr_triangle_sampleID.pdf"), width=20, height=10)
par(cex=1, mar=c(10,0,0,0))
plot(as.dendrogram(tw_kr,"quadrat"),type="triangle")
dev.off()


pdf(file.path(outPath,"Dendro_tw_kr_triangle_spp.pdf"), width=20, height=15)
par(cex=1,mar=c(10,0,0,0))
plot(as.dendrogram(tw_kr,"species"),type="triangle")
dev.off()

```
# Image (heatmap/dendrogram set)

```{r}
colrmp<-rev(brewer.pal(6,"OrRd"))
image(tw_kr, reorder=T,col=colrmp)

# better output to a giant pdf where you can see all quadrats listed
#par(cexCol=0.2, cexRow=10)
#par(cex=0.2)
pdf(file.path(outPath,"heatmap_tw_kr.pdf"), width=20, height=20)
image(tw_kr, reorder=T, col=colrmp, margins=c(10,10))
dev.off()
```
# Match KR to megacut15

match to "res" from the tw_1 analysis (here in case you need to re-run it - copied from explore_results.Rmd)
```{r}
res <- data.frame(SampID = samples, level1 = NA, level2= NA, level3 = NA, level4 = NA, level5 = NA, level6 = NA, level7 = NA, level8= NA, level9 = NA, level10 = NA, level11 = NA, level12 = NA, level13 = NA, level14= NA, level15 = NA)

for(i in 1:15){
  res[,i+1]<-cut(tw_1, level = i)
}

f = function(x) length(unique(x))
lapply(res, FUN = f)
```

make a similar list for KR
```{r}

subcore.sel<-sub.coredata%>%
  data.frame()%>%
  select(which(!is.na(colSums(.))))
  
sites.sel<-rownames_to_column(subcore.sel, "sites.sel") %>%
  select(sites.sel)





resKR<-data.frame(SampID = sites.sel, level1 = NA, level2= NA, level3 = NA, level4 = NA, level5 = NA, level6 = NA, level7 = NA, level8= NA, level9 = NA, level10 = NA, level11 = NA, level12 = NA, level13 = NA, level14= NA, level15 = NA)
 
 for(i in 1:15){
   resKR[,i+1]<-cut(tw_kr, level = i)
 }
 
 f = function(x) length(unique(x))
 lapply(resKR, FUN = f)
```

match res and resKR
```{r}
res.match<-left_join(resKR, res, by = c("sites.sel"="SampID"))
```
