==> 4a_KR_megaTwinDeepSummaries_KR.Rmd <==
---
title: "megaTwin deep summaries - R Notebook"
output: html_notebook
---
As even splitting into smaller twintables can result in some being too big to digest, we want to try and summarise groups at different levels for their average pseudospp., n, max pseudospp? to be better able to compare sub-branches.

#library
```{r}
library(twinspan)
library(dplyr, quietly = TRUE)
library(tibble)
library(Matrix)
library(rgdal)
```

# Load environmental data for later

```{r}
envK <- read.csv(file.path(dataPath,"KF50_samples_env.csv"))
envR <- read.csv(file.path(dataPath,"RF50_samples_env.csv"))

colnames(envK)
colnames(envR)
```

# Create matrix
we need the twintable matrix to work with, which we can get with twin2mat and this needs to be merged with resKR
```{r}
tw_kr.df<-rownames_to_column(as.data.frame(twin2mat(tw_kr)))


#make res again if need be
resKR<-data.frame(SampID = sites.sel, level1 = NA, level2= NA, level3 = NA, level4 = NA, level5 = NA, level6 = NA, level7 = NA, level8= NA, level9 = NA, level10 = NA, level11 = NA, level12 = NA, level13 = NA, level14= NA, level15 = NA)
 
 for(i in 1:15){
   resKR[,i+1]<-cut(tw_kr, level = i)
 }
 
 f = function(x) length(unique(x))
 lapply(resKR, FUN = f)
 
#join together
res.matKR<-left_join(resKR, tw_kr.df, by = c("sites.sel"="rowname"))


rownames(res.matKR) <- res.matKR[,1]
```


# Subset to group at given cut level

```{r}
# Set number of classes needed
n<-7

# Make reclassification table
# reclassify <- data.frame(level=c(1,4,4,4,4,3,3),
#                          class=c(3,18,19,16,17,11,10),
#                          new_class=sprintf("KR%s",seq(1:n)))

reclassify <- data.frame(level=c(2,4,4,4,4,2,2),
                         class=c(7,27,26,24,25,5,4),
                         new_class=sprintf("KR%s",seq(1:n)))


res.matKRREF<-res.matKR


# Reclassify
for(i in 1:n){
  res.matKR[getsamplesingrp(res.matKRREF,reclassify$level[i],reclassify$class[i]),"class"]<-as.character(reclassify$new_class[i])
}

table(res.matKR$class)

my.subs<-res.matKR
#my.subs<-subset(res.mat,level8 ==298) #change as needed

```
<!-- #get pivots for parent groups -->

<!-- ```{r} -->
<!-- #parent groups -->
<!-- piv.meta.a<-my.subs %>% -->
<!--   group_by(class)%>% # change to level you want to see groups at -->
<!--   summarise(across(c(2:15),mean)) %>% -->
<!--   t()%>% -->
<!--   `colnames<-`(.[1, ])%>% -->
<!--   .[-1, ] -->

<!-- ``` -->
<!-- --- -->
<!-- # DO NOT USE YET -->
<!-- # OPTIONAL  add newdata -->
<!-- If there are any samples with poor vis or that were too short that were excluded from the analysis they can be added back in here, with their level15 classes predicted, then all subsequent summaries can include their data -->

<!-- ## predict -->

<!-- The 41 excluded samples will be treated as new data to see if they fit into a group. As 3 excluded samples had issues (dive aborted, "where is it?") only the 38 in fj.newfilt.out can be used (object made in kongsRijpTWINSPAN.Rmd) -->

<!-- ## make new data matrix and predict -->

<!-- NB NOT READY - NEED PSEUDOSPECIES VALUES IN TABLE RATHER THAN REAL DENSITIY VALUES - NEEDS RECODING -->

<!-- ```{r} -->
<!-- new.data<-trimmedwide %>% -->
<!--   filter(SampID %in% fj.newfilt.out$SampID) -->

<!-- sites.sel.nd<-trimmedwide %>% -->
<!--   filter(SampID %in% fj.newfilt.out$SampID)  %>% -->
<!--   select(SampID) -->

<!-- level15<-predict(tw_kr, new.data) -->
<!-- group.df<-as.data.frame(level15) -->


<!-- #get cutlevels -->
<!-- new.data.pseu<-twinsform(new.data[,-1], cutlevels=cutlevelsPaal) -->

<!-- new.data.gp<-cbind(sites.sel.nd,new.data.pseu, group.df) -->
<!-- names(new.data.gp)[names(new.data.gp) == 'sites.sel.nd'] <- 'sites.sel'  -->

<!-- ``` -->

<!-- ## find biotope class for group (lev15) -->
<!-- NB this is run after the megaTwinDeepSummaries_KR which is where my.subs comes from -->
<!-- ```{r} -->

<!-- bio.lookup<-my.subs %>% -->
<!--   group_by(level15)%>% -->
<!--   summarize (classA =names(which.max(table(class))))%>% -->
<!--   select(level15, classA) -->

<!-- names(bio.lookup)[names(bio.lookup) == 'classA'] <- 'class'  -->


<!-- new.data.class<-new.data.gp %>% -->
<!--   left_join(bio.lookup, by=("level15")) -->

<!-- # however these are not psuedospecies values so we need to remake this table using the original cutoffs. -->


<!-- ``` -->
<!-- ## merge datasets -->

<!-- ```{r} -->
<!-- res.matKR<-bind_rows(res.matKR,new.data.class) -->

<!-- ``` -->

<!-- ------ -->
# START AGAIN HERE

# merge matrix with selected env var
```{r}
#select var
envK.sel<-envK %>%
  select(SampID,
           KF50_bathy_fmn3,
           KF50_slope3_fmn3,
           KF50_Sv800_SALTmin_fmn3,
           KF50_Sv800_TEMPmean_fmn3,
           KF50_Sv800_SPDmax_fmn3,
           sedClass50)
colnames(envK.sel)<-c("SampID",
           "F50_bathy_fmn3",
           "F50_slope3_fmn3",
           "F50_Sv800_SALTmin_fmn3",
           "F50_Sv800_TEMPmean_fmn3",
           "F50_Sv800_SPDmax_fmn3",
           "sedClass50")

envR.sel<-envR %>%
  select(SampID,
           KF50_bathy_fmn3,
           RF50_slope3_fmn3,
           RF50_Sv800_SALTmin_fmn3,
           RF50_Sv800_TEMPmean_fmn3,
           RF50_Sv800_SPDmax_fmn3,
           sedClass50)
colnames(envR.sel)<-c("SampID",
           "F50_bathy_fmn3",
           "F50_slope3_fmn3",
           "F50_Sv800_SALTmin_fmn3",
           "F50_Sv800_TEMPmean_fmn3",
           "F50_Sv800_SPDmax_fmn3",
           "sedClass50")

env<-envK.sel %>%
  bind_rows(envR.sel)


#merge with dataset

res.matKR<- res.matKR %>%
  left_join(., env, by=c("sites.sel"="SampID"))



rownames(res.matKR) <- res.matKR[,1]

my.subs<-res.matKR
```


# pivot at given cut level

Make table of metadata (parent groups, n sub groups at level 15, and n samples per group)

```{r}
#parent groups
piv.meta.a<-my.subs %>%
  group_by(class)%>% # change to level you want to see groups at
  summarise(across(c(2:15),mean)) %>%
  t()%>%
  `colnames<-`(.[1, ])%>%
  .[-1, ]



# n lev 15 groups
piv.meta.15<-my.subs %>%
  group_by(class)%>% # change to level you want to see groups at
  summarise(nSubGpL15 = n_distinct(level15)) %>%
  t()%>%
  as.data.frame() %>%
  `colnames<-`(.[1, ])%>%
  .[-1, ]

my.subs.noNA<-na.omit(my.subs)

piv.meta.env<-my.subs.noNA %>%
  group_by(class)%>% # change to level you want to see groups at
  #summarise(across(c(141:145),mean)) %>%
  summarise(bathy_av= mean(F50_bathy_fmn3),
            bathy_min= min(F50_bathy_fmn3),
            bathy_max= max(F50_bathy_fmn3),
            slope3_av=mean(F50_slope3_fmn3),
            slope3_min=min(F50_slope3_fmn3),
            slope3_max=max(F50_slope3_fmn3),
            salt_av=mean(F50_Sv800_SALTmin_fmn3),
            salt_min=min(F50_Sv800_SALTmin_fmn3),
            salt_max=max(F50_Sv800_SALTmin_fmn3),
            temp_av=mean(F50_Sv800_TEMPmean_fmn3),
            temp_min=min(F50_Sv800_TEMPmean_fmn3),
            temp_max=max(F50_Sv800_TEMPmean_fmn3),
            cspeed_av=mean(F50_Sv800_SPDmax_fmn3),
            cspeed_min=min(F50_Sv800_SPDmax_fmn3),
            cspeed_max=max(F50_Sv800_SPDmax_fmn3))%>%
  t()%>%
  as.data.frame() %>%
  `colnames<-`(.[1, ])%>%
  .[-1, ]

getmode <- function(v) {
   uniqv <- unique(v)
   uniqv[which.max(tabulate(match(v, uniqv)))]
}

piv.meta.sed<-my.subs.noNA %>%
  group_by(class)%>% # change to level you want to see groups at
  summarise(sedClass_mode = getmode(sedClass50),
            sedClass_n= length(unique(sedClass50))) %>%
  t()%>%
  as.data.frame() %>%
  `colnames<-`(.[1, ])%>%
  .[-1, ]


#merge trawl mark data (make ure you have run "data_prep_trawlmarks.Rmd" first) Need to do this after summarising env var as there are too many NAs in trawl marks to get goo summaries
trawlwide.cln<-trawlwide %>%
  group_by(SampID) %>%
  summarise(TrawlmarkDens=max(`Trawl mark`))


res.matKR<- res.matKR %>%
  left_join(., trawlwide.cln, by =c("sites.sel"="SampID"))

my.subs<-res.matKR

# n samples
piv.meta.g<-my.subs %>%
  group_by(class)%>% # change to level you want to see groups at
  summarise(nSamp=n_distinct(sites.sel),
            trawlm_Max=max(TrawlmarkDens)) %>%
  t()%>%
  as.data.frame() %>%
  `colnames<-`(.[1, ])%>%
  .[-1, ]


piv.meta<-rbind(piv.meta.a,piv.meta.15,piv.meta.g, piv.meta.env, piv.meta.sed)
```

Make table of average psuedospp per group
```{r}
piv.av<-my.subs %>%
  group_by(class)%>% # change to level you want to see groups at
  summarise(across(c(17:133),mean)) %>%
  t()%>%
 # na.omit()%>%
  `colnames<-`(.[1, ])%>%
  .[-1, ] %>%
  type.convert()
#piv.av[,] <- sapply(piv.av[,], as.numeric)


round_df <- function(x, digits) {
    # round all numeric variables
    # x: data frame 
    # digits: number of digits to round
    numeric_columns <- sapply(x, mode) == 'numeric'
    x[numeric_columns] <-  round(x[numeric_columns], digits)
    x
}

piv.av<-round_df(piv.av, 1)
```
Make table of proportion of n spp obs (non zero pseudospp)/n samples per group

```{r}
piv.n<-my.subs %>%
  group_by(class)%>% # change to level you want to see groups at
  summarise(across(c(17:133),nnzero)) %>%
  t()%>%
  #na.omit()%>%
  as.data.frame() %>%
  `colnames<-`(.[1, ])%>%
  .[-1, ] %>%
  type.convert()

piv.tot <- sapply(piv.meta.g[1,], as.numeric)
#piv.n[,] <- sapply(piv.n[,], as.numeric)

piv.prop<-mapply('/',piv.n,piv.tot)
rownames(piv.prop)<-rownames(piv.n)
piv.prop<-round_df(piv.prop,2)

```
Make table of maximum pseudospp per group
```{r}
piv.mx<-my.subs %>%
  group_by(class)%>% # change to level you want to see groups at
  summarise(across(c(17:133),max)) %>%
  t()%>%
 # na.omit()%>%
  `colnames<-`(.[1, ])%>%
  .[-1, ]
```


cbind av pseudospp, evenness, max pseudospp tables

```{r}
out.piv<-cbind(piv.av, piv.prop, piv.mx)
values<-c(rep("av",ncol(piv.av)), rep("prop",ncol(piv.prop)), rep("max", ncol(piv.mx)))

out.piv<-rbind(values, out.piv)
```


#output csvs

```{r}
# write.csv(piv.av, 
#           file = file.path(outPath, ("V6_av_pseudoSpp.csv"))) #change label
# 
# write.csv(piv.prop, 
#           file = file.path(outPath, ("V6_prop_obs.csv"))) #change label

write.csv(out.piv, 
          file = file.path(outPath, ("KR_FINAL_av_proportion_max-10.csv"))) #change label

write.csv(piv.meta, 
          file = file.path(outPath, ("KR_FINAL_metadata-10.csv"))) #change label

```





#output list of all stations with classes
This means merging with the excluded stations and retaining only the SampID and the biotope class.

##import (if not done above) and predict to excluded samples
```{r}
new.data<-trimmedwide %>%
  filter(SampID %in% fj.newfilt.out$SampID)

#sites.sel.nd<-trimmedwide %>%
 # filter(SampID %in% fj.newfilt.out$SampID)  %>%
 # select(SampID)

level15<-predict(tw_kr, new.data)
group.df<-as.data.frame(level15)


new.data.gp<-cbind(new.data, group.df)#,sites.sel.nd)
#names(new.data.gp)[names(new.data.gp) == 'sites.sel.nd'] <- 'sites.sel'

```

## find biotope class for group (lev15)
NB this is run after the megaTwinDeepSummaries_KR which is where my.subs comes from
```{r}

bio.lookup<-my.subs %>%
  group_by(level15)%>%
  summarize (classA =names(which.max(table(class))))%>%
  select(level15, classA)

names(bio.lookup)[names(bio.lookup) == 'classA'] <- 'class'


new.data.class<-new.data.gp %>%
  left_join(bio.lookup, by=("level15")) #nb inc density vals not pseudospp
 

  
```
## merge datasets

```{r}

IDclass.nd<-new.data.class %>%
  select(SampID, class)

IDclass.main<-res.matKR %>%
  select(sites.sel, class) %>%
  rename(SampID=sites.sel)


sites.class<-bind_rows(IDclass.main,IDclass.nd)

```


# output classified shapefile
```{r}
# get positional info
pos<-sample_info %>%
  select(SampID,
         x_coord,
         y_coord,
         VL)


sites.class<-sites.class%>%
  left_join(pos)


sites.class.shp <- SpatialPointsDataFrame(coords = cbind(sites.class$x_coord,
                                                     sites.class$y_coord),
                                      proj4string = CRS("+init=epsg:32633"),
                                      data =  sites.class)

writeOGR(sites.class.shp, dsn = file.path(outPath, "Spatial"), "KR_classified-10_corrected", driver = "ESRI Shapefile", overwrite_layer = TRUE)

```
with R number
```{r}
sites.class.R<-sites.class%>%
  left_join(refer, by=c("VL"="sample_no"))%>%
  select(SampID,
         x_coord,
         y_coord,
         VL,
         refstation_no)
write.csv(sites.class.R, 
          file = file.path(outPath, ("KR_sample_class_Rstation.csv")))
```


<!-- ------ -->

<!-- ## Extra checks -->
<!-- ```{r} -->
<!-- subsKr3<-subset(my.subs,class=="KR3") -->
<!-- subsKr4<-subset(my.subs,class=="KR4") -->

<!-- subsKr34<-rbind(subsKr3,subsKr4) -->

<!-- subsKr1<-subset(my.subs,class=="KR1") -->
<!-- subsKr2<-subset(my.subs,class=="KR2") -->
<!-- subsKr5<-subset(my.subs,class=="KR5") -->
<!-- subsKr6<-subset(my.subs,class=="KR6") -->
<!-- subsKr7<-subset(my.subs,class=="KR7") -->

<!-- ``` -->


<!-- ## CHECK CLASS SAMPLE LISTS -->

<!-- ```{r} -->
<!-- chk.cls.samp<-my.subs %>% -->
<!--   select(c(sites.sel, class)) -->
<!-- ``` -->

## check rogue sample
2119_B moved group - what species are in there?

```{r}

rogue<-my.subs %>%
  filter(sites.sel=="2119_B")%>%
  select (c(17:133))%>%
  t()
```

check vs mega
```{r}
sites.class.mega<-sites.class%>%
  left_join(res.match, by=c("SampID"="sites.sel")) %>%
  select(SampID, class, x_coord,y_coord,VL,level4.y,level5.y,level6.y,level7.y,level8.y,level9.y,level10.y,level11.y,level12.y,level13.y)

write.csv(sites.class.mega, 
          file = file.path(outPath, ("KR_class_megaLevel.csv"))) #change label
```
