==> 4b_classes_vs_env.Rmd <==
---
title: "classes vs environment - R Notebook"
output: html_notebook
---

For now envrionmental data was obtained by outputting the sample info from 4a_megaTwinDeepSummariesVclasses with the P class info attached, and using arcGIS to sample the rasters used in the seaPen study. This can be redone with the best available data after Margaret has collected it and the sampling can be performed in R.

#Libraries
```{r}
library(ggplot2)
library(raster)
library(rgdal)
library(tibble)
library(dplyr)
library(ggbiplot)
```
#assemble env var and classes/samples


```{r}
samp.clss<-my.subs %>%
  dplyr::select(SampID, class, level1, level2, level3, level4, level5, level6, level7, level8, level9, level10, level11, level12, level13, level14, level15)


```

# Load Data Env Data (if not already in environment: samp.EnvVar)
```{r}
##If not already loaded you can load the environmental data per sample file that was output in 2b_data_prep_envVarPerSamp.Rmd
# samp.envVar <- read.csv(file.path(outPath, ("samp_envVar.csv")))
```

#Merge class list w samp.EnvVar and add some new SedClasses
```{r}

soft.sed<-c("mud", "clay", "sandy_mud")
coarse.sed<-c("sand", "shellsand","sand_with_shells","sand_with_biomaterial", "clay_spicules")
hard.sed<-c("pebble","coral_rubble", "lophelia_rubble")
hard.solid<-c("bedrock", "boulder", "cobble", "karbonatskorpe","compacted_sediment", "consolidated_clay")

# order columns as want them later
samp.envVar<-samp.envVar %>%
  select("SampID",
         "mean_depth",
         "x_coord","y_coord",
         "salt_max..tif", "salt_mean..tif", "salt_min..tif",  "salt_std..tif",  
         "spd_max..tif","spd_mean..tif","spd_min..tif","spd_std..tif","u_bott_mean..tif","v_bott_mean..tif",   
         "temp_max..tif", "temp_mean..tif", "temp_min..tif",  "temp_std..tif",  
         "bedrock","boulder","cobble","karbonatskorpe","compacted_sediment","consolidated_clay",
         "clay_(outcropping)",
         "mud","sandy_mud", "clay",
         "pebble","coral_rubble", "lophelia_rubble", 
         "dead_lophelia", "live_lophelia",  
         "sand","shellsand","sand_with_shells","sand_with_biomaterial", 
         "clay_spicules", "gelatinous_snow", "biobush")


class.env<-samp.clss%>%
  left_join(samp.envVar)%>%
  rowwise()%>%
  dplyr::mutate(softSed=sum(c_across(any_of(soft.sed)), na.rm = T), 
  coarseSed=sum(c_across(any_of(coarse.sed)), na.rm = T),
  hardSed=sum(c_across(any_of(hard.sed)), na.rm = T),
  hardSolid=sum(c_across(any_of(hard.solid)), na.rm = T))%>%#not working well - bedrock 100 in class 3 has come out as sum hard 5.9...?
  ungroup()
```






#reorder levels of class to be in order

```{r}
#class.ls<-c("P1","P2","P3","P4","P5","P6","P7","P8","P9","P10",
#,     "P11","P12","P13","P14","P15","P16","P17","P18","P19","P26",
#,     "P20","P21","P22","P23","P24","P25")

# class.ls<-c("P1","P2","P3","P4","P5","P6","P7","P8","P9","P10",
#,      "P11","P12","P13","P14","P15","P16","P17","P18","P19","P20",
#,      "P21","P22","P23","P24","P25","P26","P27","P28","P29","P30",
#,      "P31","P32","P33","P34","P35","P36","P37","P38","P39","P40",
#,      "P41","P42","P43","P44","P45","P46","P47","P48","P49","P50",
#,      "P51","P52","P53","P54","P55","P56","P57","P58","P59","P60",
#,      "P61","P62","P63","P64")


# class.ls<-c("SubTmp1","SubTmp2","SubTmp3","SubTmp4","SubTmp5","SubTmp6","SubTmp7","SubTmp8","SubTmp9","SubTmp10",
#       "SubTmp11","SubTmp12","SubTmp13","SubTmp14","SubTmp15","SubTmp16","SubTmp17","SubTmp18","SubTmp19","SubTmp20",
#       "SubTmp21","SubTmp22","SubTmp23","SubTmp24","SubTmp25","SubTmp26","SubTmp27","SubTmp28","SubTmp29","SubTmp30",
#       "SubTmp31","SubTmp32","SubTmp33","SubTmp34","SubTmp35","SubTmp36","SubTmp37","SubTmp38","SubTmp39","SubTmp40",
#       "SubTmp41","SubTmp42","SubTmp43","SubTmp44","SubTmp45","SubTmp46","SubTmp47","SubTmp48","SubTmp49","SubTmp50",
#       "SubTmp51","SubTmp52","SubTmp53","SubTmp54","SubTmp55","SubTmp56","SubTmp57","SubTmp58","SubTmp59","SubTmp60",
#       "SubTmp61","SubTmp62","SubTmp63","SubTmp64","SubTmp65","SubTmp66","SubTmp67","SubTmp68","SubTmp69","SubTmp70",
#       "SubTmp71","SubTmp72","SubTmp73","SubTmp74")
# 
# class.env$class <- factor(class.env$class, levels = class.ls)

class.env$level1<-as.factor(class.env$level1)
class.env$level2<-as.factor(class.env$level2)
class.env$level3<-as.factor(class.env$level3)
class.env$level4<-as.factor(class.env$level4)
class.env$level5<-as.factor(class.env$level5)
class.env$level6<-as.factor(class.env$level6)
class.env$level7<-as.factor(class.env$level7)
class.env$level8<-as.factor(class.env$level8)
class.env$level9<-as.factor(class.env$level9)
class.env$level10<-as.factor(class.env$level10)
class.env$level11<-as.factor(class.env$level11)
class.env$level12<-as.factor(class.env$level12)
class.env$level13<-as.factor(class.env$level13)
class.env$level14<-as.factor(class.env$level14)
class.env$level15<-as.factor(class.env$level15)

class.env$mean_depth<-class.env$mean_depth*-1

class.env[class.env==-9999]<-NA

summary(class.env)

summary(class.env$class)

colnames(class.env)

```




#Plots

Depth (Bathy)
```{r}
png("U:/Mareano/VIDEOLAB/VIDEO DATA/200m_scale_species_by_sample/Data_Delivery_2021/outputs/TMP class 0122/SubTMPMrg_env_bathy.png",width = 1200, height = 500, units = "px")
ggplot(class.env, aes(x=class, y=mean_depth))+
  geom_boxplot()+
  theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1))
dev.off()
```
Temp Av (T_Av)
```{r}
png("U:/Mareano/VIDEOLAB/VIDEO DATA/200m_scale_species_by_sample/Data_Delivery_2021/outputs/TMP class 0122/SubTMPMrg_env_T_Av.png",width = 1200, height = 500, units = "px")
ggplot(class.env, aes(x=class, y=temp_mean..tif))+
  geom_boxplot()+   theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1))
dev.off()
```
Temp Min (T_Min)
```{r}
png("U:/Mareano/VIDEOLAB/VIDEO DATA/200m_scale_species_by_sample/Data_Delivery_2021/outputs/TMP class 0122/SubTMPMrg_env_T_Min.png",width = 1200, height = 500, units = "px")
ggplot(class.env, aes(x=class, y=temp_min..tif))+
  geom_boxplot()+   theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1))
dev.off()
```
Temp SD
```{r}
png("U:/Mareano/VIDEOLAB/VIDEO DATA/200m_scale_species_by_sample/Data_Delivery_2021/outputs/TMP class 0122/SubTMPMrg_env_T_SD.png",width = 1200, height = 500, units = "px")
ggplot(class.env, aes(x=class, y=temp_std..tif))+
  geom_boxplot()+   theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1))
dev.off()
```


Salt Av(Salt_Av)
```{r}
png("U:/Mareano/VIDEOLAB/VIDEO DATA/200m_scale_species_by_sample/Data_Delivery_2021/outputs/TMP class 0122/SubTMPMrg_env_Salt_Av.png",width = 1200, height = 500, units = "px")
ggplot(class.env, aes(x=class, y=salt_mean..tif))+
  geom_boxplot()+   theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1))
dev.off()
```

Salt Min (Salt_Min)
```{r}
png("U:/Mareano/VIDEOLAB/VIDEO DATA/200m_scale_species_by_sample/Data_Delivery_2021/outputs/TMP class 0122/SubTMPMrg_env_Salt_Min.png",width = 1200, height = 500, units = "px")
ggplot(class.env, aes(x=class, y=salt_min..tif))+
  geom_boxplot()+   theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1))
dev.off()
```
Salt SD 
```{r}
png("U:/Mareano/VIDEOLAB/VIDEO DATA/200m_scale_species_by_sample/Data_Delivery_2021/outputs/TMP class 0122/SubTMPMrg_env_Salt_SD.png",width = 1200, height = 500, units = "px")
ggplot(class.env, aes(x=class, y=salt_std..tif))+
  geom_boxplot()+   theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1))
dev.off()
```

Cspd_Av
```{r}
png("U:/Mareano/VIDEOLAB/VIDEO DATA/200m_scale_species_by_sample/Data_Delivery_2021/outputs/TMP class 0122/SubTMPMrg_env_Cspd_Av.png",width = 1200, height = 500, units = "px")
ggplot(class.env, aes(x=class, y=spd_mean..tif))+
  geom_boxplot()+   theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1))
dev.off()
```

Cspd_Mx
```{r}
png("U:/Mareano/VIDEOLAB/VIDEO DATA/200m_scale_species_by_sample/Data_Delivery_2021/outputs/TMP class 0122/SubTMPMrg_env_Cspd_Mx.png",width = 1200, height = 500, units = "px")
ggplot(class.env, aes(x=class, y=spd_max..tif))+
  geom_boxplot()+   theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1))
dev.off()
```
Cspd_Mx
```{r}
png("U:/Mareano/VIDEOLAB/VIDEO DATA/200m_scale_species_by_sample/Data_Delivery_2021/outputs/TMP class 0122/SubTMPMrg_env_Cspd_SD.png",width = 1200, height = 500, units = "px")
ggplot(class.env, aes(x=class, y=spd_std..tif))+
  geom_boxplot()+   theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1))
dev.off()
```

Xcoord
```{r}
png("U:/Mareano/VIDEOLAB/VIDEO DATA/200m_scale_species_by_sample/Data_Delivery_2021/outputs/TMP class 0122/SubTMPMrg_env_Xcoord.png",width = 1200, height = 500, units = "px")
ggplot(class.env, aes(x=class, y=x_coord))+
  geom_boxplot()+   theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1))
dev.off()
```

Ycoord
```{r}
png("U:/Mareano/VIDEOLAB/VIDEO DATA/200m_scale_species_by_sample/Data_Delivery_2021/outputs/TMP class 0122/SubTMPMrg_env_Ycoord.png",width = 1200, height = 500, units = "px")
ggplot(class.env, aes(x=class, y=y_coord))+
  geom_boxplot()+   theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1))
dev.off()
```

# SEDIMENTS

Boulders
```{r}
png("U:/Mareano/VIDEOLAB/VIDEO DATA/200m_scale_species_by_sample/Data_Delivery_2021/outputs/TMP class 0122/SubTMPMrg_env_Sed_Bould.png",width = 1200, height = 500, units = "px")
ggplot(class.env, aes(x=class, y=boulder))+
  geom_boxplot()+   theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1))
dev.off()
```
Cobbles
```{r}
png("U:/Mareano/VIDEOLAB/VIDEO DATA/200m_scale_species_by_sample/Data_Delivery_2021/outputs/TMP class 0122/SubTMPMrg_env_Sed_Cobb.png",width = 1200, height = 500, units = "px")
ggplot(class.env, aes(x=class, y=cobble))+
  geom_boxplot()+   theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1))
dev.off()
``` 
Pebbles
```{r}
png("U:/Mareano/VIDEOLAB/VIDEO DATA/200m_scale_species_by_sample/Data_Delivery_2021/outputs/TMP class 0122/SubTMPMrg_env_Sed_Pebb.png",width = 1200, height = 500, units = "px")
ggplot(class.env, aes(x=class, y=pebble))+
  geom_boxplot()+   theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1))
dev.off()
``` 
Sandy_mud
```{r}
png("U:/Mareano/VIDEOLAB/VIDEO DATA/200m_scale_species_by_sample/Data_Delivery_2021/outputs/TMP class 0122/SubTMPMrg_env_Sed_SandyMud.png",width = 1200, height = 500, units = "px")
ggplot(class.env, aes(x=class, y=sandy_mud))+
  geom_boxplot()+   theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1))
dev.off()
``` 

SedSand (Sed_Sand)
```{r}
png("U:/Mareano/VIDEOLAB/VIDEO DATA/200m_scale_species_by_sample/Data_Delivery_2021/outputs/TMP class 0122/SubTMPMrg_env_Sed_Sand.png",width = 1200, height = 500, units = "px")
ggplot(class.env, aes(x=class, y=sand))+
  geom_boxplot()+   theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1))
dev.off()
```

SedMud
```{r}
png("U:/Mareano/VIDEOLAB/VIDEO DATA/200m_scale_species_by_sample/Data_Delivery_2021/outputs/TMP class 0122/SubTMPMrg_env_Sed_Mud.png",width = 1200, height = 500, units = "px")
ggplot(class.env, aes(x=class, y=mud))+
  geom_boxplot()+   theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1))
dev.off()
```
Bedrock
```{r}
png("U:/Mareano/VIDEOLAB/VIDEO DATA/200m_scale_species_by_sample/Data_Delivery_2021/outputs/TMP class 0122/SubTMPMrg_env_Sed_Bedrock.png",width = 1200, height = 500, units = "px")
ggplot(class.env, aes(x=class, y=bedrock))+
  geom_boxplot()+   theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1))
dev.off()
```
Compacted Sediment
```{r}
png("U:/Mareano/VIDEOLAB/VIDEO DATA/200m_scale_species_by_sample/Data_Delivery_2021/outputs/TMP class 0122/SubTMPMrg_env_Sed_CompSed.png",width = 1200, height = 500, units = "px")
ggplot(class.env, aes(x=class, y=compacted_sediment))+
  geom_boxplot()+   theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1))
dev.off()
```
shellsand
```{r}
png("U:/Mareano/VIDEOLAB/VIDEO DATA/200m_scale_species_by_sample/Data_Delivery_2021/outputs/TMP class 0122/SubTMPMrg_env_Sed_Shellsand.png",width = 1200, height = 500, units = "px")
ggplot(class.env, aes(x=class, y=shellsand))+
  geom_boxplot()+   theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1))
dev.off()
```
clay
```{r}
png("U:/Mareano/VIDEOLAB/VIDEO DATA/200m_scale_species_by_sample/Data_Delivery_2021/outputs/TMP class 0122/SubTMPMrg_env_Sed_Clay.png",width = 1200, height = 500, units = "px")
ggplot(class.env, aes(x=class, y=clay))+
  geom_boxplot()+   theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1))
dev.off()
```
consolidated_clay
```{r}
png("U:/Mareano/VIDEOLAB/VIDEO DATA/200m_scale_species_by_sample/Data_Delivery_2021/outputs/TMP class 0122/SubTMPMrg_env_Sed_ConsolClay.png",width = 1200, height = 500, units = "px")
ggplot(class.env, aes(x=class, y=consolidated_clay))+
  geom_boxplot()+   theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1))
dev.off()
```
clay outcropping
```{r}
png("U:/Mareano/VIDEOLAB/VIDEO DATA/200m_scale_species_by_sample/Data_Delivery_2021/outputs/TMP class 0122/SubTMPMrg_env_Sed_ClayOutcrop.png",width = 1200, height = 500, units = "px")
ggplot(class.env, aes(x=class, y="clay_(outcropping)"))+
  geom_boxplot()+   theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1))
dev.off()
```

karbonatskorpe
```{r}
png("U:/Mareano/VIDEOLAB/VIDEO DATA/200m_scale_species_by_sample/Data_Delivery_2021/outputs/TMP class 0122/SubTMPMrg_env_Sed_KarbonatSk.png",width = 1200, height = 500, units = "px")
ggplot(class.env, aes(x=class, y=karbonatskorpe))+
  geom_boxplot()+   theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1))
dev.off()
```
biobush
```{r}
png("U:/Mareano/VIDEOLAB/VIDEO DATA/200m_scale_species_by_sample/Data_Delivery_2021/outputs/TMP class 0122/SubTMPMrg_env_Sed_Biobush.png",width = 1200, height = 500, units = "px")
ggplot(class.env, aes(x=class, y=biobush))+
  geom_boxplot()+   theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1))
dev.off()
```
clay_spicules
```{r}
png("U:/Mareano/VIDEOLAB/VIDEO DATA/200m_scale_species_by_sample/Data_Delivery_2021/outputs/TMP class 0122/SubTMPMrg_env_Sed_ClaySpicules.png",width = 1200, height = 500, units = "px")
ggplot(class.env, aes(x=class, y=clay_spicules))+
  geom_boxplot()+   theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1))
dev.off()
```
gelatinous_snow
```{r}
png("U:/Mareano/VIDEOLAB/VIDEO DATA/200m_scale_species_by_sample/Data_Delivery_2021/outputs/TMP class 0122/SubTMPMrg_env_Sed_GelSnow.png",width = 1200, height = 500, units = "px")
ggplot(class.env, aes(x=class, y=gelatinous_snow))+
  geom_boxplot()+   theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1))
dev.off()
```
sand with shells
```{r}
png("U:/Mareano/VIDEOLAB/VIDEO DATA/200m_scale_species_by_sample/Data_Delivery_2021/outputs/TMP class 0122/SubTMPMrg_env_Sed_SandWShells.png",width = 1200, height = 500, units = "px")
ggplot(class.env, aes(x=class, y=sand_with_shells))+
  geom_boxplot()+   theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1))
dev.off()
```
sand with biomaterial
```{r}
png("U:/Mareano/VIDEOLAB/VIDEO DATA/200m_scale_species_by_sample/Data_Delivery_2021/outputs/TMP class 0122/SubTMPMrg_env_Sed_SandWBiomateral.png",width = 1200, height = 500, units = "px")
ggplot(class.env, aes(x=class, y=sand_with_biomaterial))+
  geom_boxplot()+   theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1))
dev.off()
```






#Coral/Lophelia

Coral Rubble
```{r}
png("U:/Mareano/VIDEOLAB/VIDEO DATA/200m_scale_species_by_sample/Data_Delivery_2021/outputs/TMP class 0122/SubTMPMrg_env_Sed_CoralRubb.png",width = 1200, height = 500, units = "px")
ggplot(class.env, aes(x=class, y=coral_rubble))+
  geom_boxplot()+   theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1))
dev.off()
```
Lophelia Rubble
```{r}
png("U:/Mareano/VIDEOLAB/VIDEO DATA/200m_scale_species_by_sample/Data_Delivery_2021/outputs/TMP class 0122/SubTMPMrg_env_Sed_LopheliaRubb.png",width = 1200, height = 500, units = "px")
ggplot(class.env, aes(x=class, y=lophelia_rubble))+
  geom_boxplot()+   theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1))
dev.off()
```
Live Lophelia
```{r}
png("U:/Mareano/VIDEOLAB/VIDEO DATA/200m_scale_species_by_sample/Data_Delivery_2021/outputs/TMP class 0122/SubTMPMrg_env_Sed_LiveLophelia.png",width = 1200, height = 500, units = "px")
ggplot(class.env, aes(x=class, y=live_lophelia))+
  geom_boxplot()+   theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1))
dev.off()
```
Live Lophelia
```{r}
png("U:/Mareano/VIDEOLAB/VIDEO DATA/200m_scale_species_by_sample/Data_Delivery_2021/outputs/TMP class 0122/SubTMPMrg_env_Sed_LiveLophelia.png",width = 1200, height = 500, units = "px")
ggplot(class.env, aes(x=class, y=live_lophelia))+
  geom_boxplot()+   theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1))
dev.off()
```

Dead Lophelia
```{r}
png("U:/Mareano/VIDEOLAB/VIDEO DATA/200m_scale_species_by_sample/Data_Delivery_2021/outputs/TMP class 0122/SubTMPMrg_env_Sed_DeadLophelia.png",width = 1200, height = 500, units = "px")
ggplot(class.env, aes(x=class, y=dead_lophelia))+
  geom_boxplot()+   theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1))
dev.off()
```


#Bex new sedClasses
NOT WORKING WELL YET - BEDROCK IN CLASS 3 HAD 100 IN THERE BUT SUM WAS >6, DOING LONGHAND IN EXCEL NLY FOR MOMENT - DEAL WITH LATER
<!-- soft sed -->
<!-- ```{r} -->
<!-- png("U:/Mareano/VIDEOLAB/VIDEO DATA/200m_scale_species_by_sample/Data_Delivery_2021/outputs/TMP class 0122/SubTMPMrg_env_Sed_SOFT.png",width = 1200, height = 500, units = "px") -->
<!-- ggplot(class.env, aes(x=class, y=softSed))+ -->
<!--   geom_boxplot()+   theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1)) -->
<!-- dev.off() -->
<!-- ``` -->
<!-- coarse sed -->
<!-- ```{r} -->
<!-- png("U:/Mareano/VIDEOLAB/VIDEO DATA/200m_scale_species_by_sample/Data_Delivery_2021/outputs/TMP class 0122/SubTMPMrg_env_Sed_COARSE.png",width = 1200, height = 500, units = "px") -->
<!-- ggplot(class.env, aes(x=class, y=coarseSed))+ -->
<!--   geom_boxplot()+   theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1)) -->
<!-- dev.off() -->
<!-- ``` -->
<!-- hard sed -->
<!-- ```{r} -->
<!-- png("U:/Mareano/VIDEOLAB/VIDEO DATA/200m_scale_species_by_sample/Data_Delivery_2021/outputs/TMP class 0122/SubTMPMrg_env_Sed_HARD_SED.png",width = 1200, height = 500, units = "px") -->
<!-- ggplot(class.env, aes(x=class, y=hardSed))+ -->
<!--   geom_boxplot()+   theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1)) -->
<!-- dev.off() -->
<!-- ``` -->
<!-- hard solid -->
<!-- ```{r} -->
<!-- png("U:/Mareano/VIDEOLAB/VIDEO DATA/200m_scale_species_by_sample/Data_Delivery_2021/outputs/TMP class 0122/SubTMPMrg_env_Sed_HARD_SOLID.png",width = 1200, height = 500, units = "px") -->
<!-- ggplot(class.env, aes(x=class, y=hardSolid))+ -->
<!--   geom_boxplot()+   theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1)) -->
<!-- dev.off() -->
<!-- ``` -->
<!-- # OUTPUT averages per class -->

```{r}
av.class.env<-class.env%>%
  group_by(class)%>%
  summarise_all(mean, na.rm=T)%>%
  select(-c("SampID"))%>%
  t()%>%
  as.data.frame() %>%
  `colnames<-`(.[1, ])%>%
  .[-(1:15), ]%>%
  type.convert()
  

write.csv(av.class.env, file = file.path(outPath, ("SubTMPMrgclass_envVar_summaries.csv")))
```
