==> 5_make_deliverable_table.Rmd <==
---
title: "Fagligforum product"
output: html_notebook
---

## Libraries
```{r}
library(dplyr)
library(RPostgreSQL)
#library(twinspan)
```

## Create table
```{r}
sheet1 <- data.frame(StationNr = NA,
                     SampleNr = unique(spmatrix$SampID),
                     x_coord = NA,
                     y_coord = NA,
                     Biotope_class = NA,
                     PredictedYN = NA,
                     Region = NA)

sheet1 <- sheet1 %>%
  mutate(StationNr = as.integer(sample_info$VL[match(SampleNr, sample_info$SampID)])) %>%
  mutate(x_coord = sample_info$x_coord[match(SampleNr, sample_info$SampID)]) %>%
  mutate(y_coord = sample_info$y_coord[match(SampleNr, sample_info$SampID)]) %>%
  arrange(StationNr) 
```

## Import station metadata from Marbunn
This bit can replace the chunk where we import reference table and create the "refer" object. Safer to do it this way.
```{r}
postgres_driver <- dbDriver("PostgreSQL")
postgres_conn <- dbConnect(postgres_driver,
                           dbname = "marbunn",
                           host = "postgres9.imr.no",
                           port = 5432,
                           user = "marbunn",
                           password = "89kk.KoqA")
query_res <- dbGetQuery(postgres_conn, "select * from tab_station where equipment = 2")

```

## Add data from reference table
```{r}
sheet1 <- sheet1 %>%
  mutate(Region = query_res$location[match(StationNr, query_res$refstation_no)])
```


## Predict class for excluded samples
Also, make a lookup table to pull our label
```{r}
# new.data <- left_join(excl.samp,trimmedwide)
# new.data <- new.data %>%
#   select(-colnames(excl.samp)[-1]) %>%
#   filter(!is.na(aplysillasulfurea)) # just using first species (if that is NA, all are NA)

# level15<-predict(tw_1, new.data)
# group.df<-as.data.frame(level15)

# new.data.gp<-cbind(new.data$SampID, group.df)
# colnames(new.data.gp)[1]<-"SampID"

# rec <- samp.clss %>% group_by(level15) %>% summarize(class = first(class))
```

## Add biotope class data to sheet 1
```{r}
samp.class <- my.subs.small

sheet1 <- sheet1 %>%
  mutate(Biotope_class=case_when(
    SampleNr %in% row.names(coredata)~ samp.clss$class[match(SampleNr,samp.clss$SampID)],
    #SampleNr %in% new.data$SampID ~ rec$class[match(new.data.gp$level15[match(SampleNr,new.data.gp$SampID)],rec$level15)]
  ))

#sheet1 <- sheet1 %>%
#  mutate(PredictedYN=case_when(
#    SampleNr %in% row.names(coredata)~ "N",
#    SampleNr %in% new.data$SampID ~ "Y"
#  ))
```

## Export
```{r}
sheet1 <- sheet1 %>%
 # transform(Biotope_class=gsub("P", "pB", Biotope_class)) %>%
#sheet1 <- sheet1 %>%  
  filter(!is.na(Biotope_class))

write.csv(sheet1, file = file.path(outPath, ("sheet1.csv")), row.names = FALSE)
```
